requirements.txt
tensorflow
ultralytics
streamlit
opencv-python
numpy
pillow
matplotlib
